+++ examples/svs/high-level/count/main.go
@@ -66,7 +66,7 @@ func main() {
 
 	syncPrefix, _ := enc.NameFromStr("/svs")
 	sourceName, _ := enc.NameFromStr(*source)
-	callback := func(source string, seqno uint, data ndn.Data) {
+	callback := func(source string, seqno uint64, data ndn.Data) {
 		if data != nil {
 			fmt.Println(source + ": " + string(data.Content().Join()))
 		} else {
